pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "google-workspace-emad"
version = "0.1.0"
description = "Google Workspace Service Broker eMAD — Gmail, Drive, Docs, Sheets, Slides, Calendar"
requires-python = ">=3.12"
dependencies = [
    "langgraph>=0.2.60",
    "langchain-core>=0.3.28",
    "langchain-openai>=0.2.14",
    "google-api-python-client>=2.166.0",
    "google-auth>=2.38.0",
    "google-auth-httplib2>=0.2.0",
    "pydantic>=2.10.0",
    "asyncpg>=0.30.0",
    "pyyaml>=6.0.2",
    "prometheus-client>=0.21.1",
]

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
google_workspace_emad = ["prompts/*.md"]
